Modificado: dirección de las páginas, manejo de errores, algunos cambios de estilo
